requirements.txt
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyarrow>=12.0.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
streamlit>=1.30.0
plotly>=5.18.0
